Generate a non-api-key-auth version by default with instructions to add in api-key-auth features
