+++ lib/stitches_norailtie.rb
@@ -12,6 +12,7 @@ require 'stitches/render_timestamps_in_iso8601_in_json'
 require 'stitches/error'
 require 'stitches/errors'
 require 'stitches/api_generator'
+require 'stitches/api_migration_generator'
 require 'stitches/add_deprecation_generator'
 require 'stitches/add_enabled_to_api_clients_generator'
 require 'stitches/add_disabled_at_to_api_clients_generator'
